Refactor DTOs and introduce PolicyType class
- Added ContractId, PersonId, and PolicyId to EditContractDto.
- Replaced Type enum in EditPolicyDto with PolicyType.Type and added Id property.
- Removed ItemListDto and updated project file to include PolicyType.cs.
- Removed PolicyId from ListItemContractDto and Phone, Age from ListItemPersonDto.
- Updated ListItemPolicyDto and UpdatePolicyDto to use PolicyType.Type instead of Type enum.
- Created new PolicyType class with an enum for policy types.
