+++ EvidencePojisteniDto/EditContractDto.cs
@@ -8,6 +8,9 @@ namespace EvidencePojisteniDto
 {
     public class EditContractDto
     {
+        public Guid ContractId { get; set; }
+        public Guid PersonId { get; set; }
+        public Guid PolicyId { get; set; }
         public DateTime ValidFrom { get; set; }
         public DateTime ValidTo { get; set; }
         public decimal Amount { get; set; }
