+++ EvidencePojisteniDto/ListItemContractDto.cs
@@ -10,7 +10,6 @@ namespace EvidencePojisteniDto
     {
         public Guid ContractId { get; set; }
         public Guid PersonId { get; set; }
-        public Guid PolicyId { get; set; }
         public DateTime ValidFrom { get; set; }
         public DateTime ValidTo { get; set; }
         public decimal Amount { get; set; }
