+++ EvidencePojisteniDto/ListItemPersonDto.cs
@@ -11,7 +11,5 @@ namespace EvidencePojisteniDto
         public Guid PersonId { get; set; }
         public string FirstName { get; set; } 
         public string LastName { get; set; }
-        public string Phone { get; set; }
-        public int Age { get; set; }
     }
 }
