+++ EvidencePojisteniDto/ListItemPolicyDto.cs
@@ -5,14 +5,7 @@ namespace EvidencePojisteniDto
     public class ListItemPolicyDto
     {
         public Guid Id { get; set; }
-        public enum Type
-        {
-            Life,
-            Health,
-            Property,
-            Vehicle,
-            Travel
-        }
+        public PolicyType.Type Type { get; set; }
         public string Name { get; set; }
     }
 }
