+++ EvidencePojisteniDto/PolicyType.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EvidencePojisteniDto
+{
+    public class PolicyType
+    {
+        public enum Type
+        {
+            Life,
+            Health,
+            Property,
+            Vehicle,
+            Travel
+        }
+    }
+}
